backend/requirements.txt
django>=5.0
djangorestframework
psycopg2-binary
django-cors-headers
celery
redis
playwright
Pillow
